Client can receive data for their account and other user
Fixes issue where users couldn't see each other's paddle while playing.
Fixed issue in the creation of the game and account page that was
causing client to not receive account statistics object from server. Can
now receive the data and display it in the views. Added a basic account
page which just displays the statistics data received from the server.