Further cleaned up CSS/templates, work more with server data
Got rid of some old stuff in the jade files that related to the original
DomoMaker code, and then updated the CSS for those pages. No more images
on the whole site missing, except I need a favicon :)

Base jade file now has a mini script contained within it to accept the
local data passed to the client by express during the render call,
storing it in a local "userdata" object which is immediately frozen. The
client now uses this to display account statistics on its account page.